pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pptx-generator"
version = "0.1.0"
description = "Professional PowerPoint presentation generator with presets, themes, and visual elements"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "python-pptx>=0.6.23",
    "pydantic>=2.0",
    "Pillow>=10.0",
    "matplotlib>=3.7",
    "click>=8.1",
    "pyyaml>=6.0",
    "markdown>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
pptx-gen = "pptx_generator.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
